requirements.txt
aiohttp
aiosignal
aniso8601
apispec
async-timeout
attrs
blinker
certifi
charset-normalizer
click
decorator
filelock
Flask==3.0.0
flask-smorest
frozenlist
gunicorn
idna
imageio
imageio-ffmpeg
itsdangerous
Jinja2
llvmlite
MarkupSafe
more-itertools
moviepy
mpmath
multidict
networkx
numba
numpy
openai
openai-whisper 
packaging
Pillow
proglog
python-dotenv
regex
requests
six
sympy
torch
tqdm
typing-extensions
urllib3
webargs
Werkzeug
yarl
